notes_backend/notes_backend/notes/serializers.py
from rest_framework import serializers

from notes_backend.notes.models import Note


class NoteSerializer(serializers.ModelSerializer):

    class Meta:
        model = Note
        fields = ['id', 'title', 'content']